Develop (#10)
* добавил модели данных User и Account

* добавлены UserService и AccountService и конфиг для default properties

* добавлены тесты для UserService и AccountService

* добавлены все операции согласно ТЗ

* добавлены listener для ввода с консоли + fix

* конфигурация проекта для использования hibernate

* переопределение классов User и Account в Entity

* переопределение сервисов и process классов согласно ТЗ

+++ src/main/java/baskakov/dev/operation/process/AccountTransferProcess.java
@@ -21,9 +21,9 @@ public class AccountTransferProcess implements ProcessorOperation {
     @Override
     public void process() {
         System.out.println("Введите идентификатор счета откуда хотите перевести средства:");
-        int fromAccId = Integer.parseInt(scanner.nextLine());
+        Long fromAccId = Long.parseLong(scanner.nextLine());
         System.out.println("Введите идентификатор счета куда хотите перевести средства:");
-        int toAccId = Integer.parseInt(scanner.nextLine());
+        Long toAccId = Long.parseLong(scanner.nextLine());
         System.out.println("Введите сумму перевода:");
         int amount = Integer.parseInt(scanner.nextLine());
         accountService.transferMoney(fromAccId, toAccId, amount);
